requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
sqlparse==0.5.3
sentence-transformers
scikit-learn
deep-translator
langdetect
python-dotenv
openai
django-cors-headers
python-decouple